refactor(ui): split repository workspace tauri services
